+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog of lizard-map
 1.64 (unreleased)
 -----------------
 
+- Added BackgroundMap and Setting models. Model BackgroundMap used to
+  store the configuration of single background maps and if the map is
+  the default one. Setting is to store global, end-user changeable
+  settings, such as startlocation_x, startlocation_y, etc.
+
 - Added twitter icon.